remove job
