--- JobSeeker/Job/Job/admin.py
@@ -1,6 +1,0 @@
-from django.contrib import admin
-from .models import JobCategory, Jobs
-
-admin.site.register(Jobs)
-admin.site.register(JobCategory)
-
