--- JobSeeker/Job/Job/test.py
@@ -1,5 +1,0 @@
-from Job.models import Jobs
-
-def fun():
-    jobs=Jobs.objects.values()
-    print(jobs)
