--- JobSeeker/requirements.txt
@@ -1,17 +1,0 @@
-asgiref==3.9.1
-certifi==2025.8.3
-charset-normalizer==3.4.3
-dj-rest-auth==7.0.1
-Django==5.2.5
-django-allauth==65.11.0
-django-cors-headers==4.7.0
-django-ratelimit==4.1.0
-djangorestframework==3.16.1
-idna==3.10
-pillow==11.3.0
-python-decouple==3.8
-requests==2.32.5
-six==1.17.0
-sqlparse==0.5.3
-tzdata==2025.2
-urllib3==2.5.0
